requirements.txt
pandas
sqlalchemy
python-dotenv
streamlit
plotly
sentence-transformers
faiss-cpu
